requirements.txt
requests
sentence-transformers
faiss-cpu
pyyaml
python-dotenv
PyPDF2
beautifulsoup4
ollama
